[RATOM-92] Revert to button
