+++ package.json
@@ -23,7 +23,8 @@
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.2.0",
     "styled-components": "^4.4.1",
-    "tinycolor2": "^1.4.1"
+    "tinycolor2": "^1.4.1",
+    "moment": "^2.24.0"
   },
   "scripts": {
     "start": "react-app-rewired start",
@@ -57,8 +58,6 @@
     "eslint-plugin-prettier": "^3.1.2",
     "eslint-plugin-react": "^7.17.0",
     "eslint-plugin-react-hooks": "^1.7.0",
-    "moment": "^2.24.0",
-    "prettier": "^1.19.1",
-    "react-moment": "^0.9.7"
+    "prettier": "^1.19.1"
   }
 }
